+++ packages/inngest/src/components/InngestCommHandler.ts
@@ -1418,7 +1418,7 @@ export class InngestCommHandler<
     // Handle opting in to optimized parallelism in v3.
     if (
       version === ExecutionVersion.V1 &&
-      fn.fn["shouldOptimizeParallelism"]()
+      fn.fn["shouldOptimizeParallelism"]?.()
     ) {
       version = ExecutionVersion.V2;
     }