+++ packages/inngest/src/components/connect/index.ts
@@ -1,10 +1,14 @@
 import { WaitGroup } from "@jpwilliams/waitgroup";
 import debug, { type Debugger } from "debug";
 import { ulid } from "ulidx";
-import { envKeys, headerKeys, queryKeys } from "../../helpers/consts.ts";
-import { allProcessEnv, getPlatformName } from "../../helpers/env.ts";
-import { parseFnData } from "../../helpers/functions.ts";
-import { hashSigningKey } from "../../helpers/strings.ts";
+import { envKeys, headerKeys, queryKeys } from "../../helpers/consts.js";
+import {
+  allProcessEnv,
+  getEnvironmentName,
+  getPlatformName,
+} from "../../helpers/env.js";
+import { parseFnData } from "../../helpers/functions.js";
+import { hashSigningKey } from "../../helpers/strings.js";
 import {
   ConnectMessage,
   type GatewayExecutorRequestData,
@@ -47,6 +51,8 @@ import {
 
 const ResponseAcknowlegeDeadline = 5_000;
 const WorkerHeartbeatInterval = 10_000;
+
+const InngestBranchEnvironmentSigningKeyPrefix = "signkey-branch-";
 
 interface connectionEstablishData {
   marshaledCapabilities: string;
@@ -102,6 +108,8 @@ class WebSocketWorkerConnection implements WorkerConnection {
   private _hashedSigningKey: string | undefined;
   private _hashedFallbackKey: string | undefined;
 
+  private _inngestEnv: string | undefined;
+
   /**
    * A set of gateways to exclude from the connection.
    */
@@ -131,8 +139,18 @@ class WebSocketWorkerConnection implements WorkerConnection {
     }
 
     this.inngest = options.apps[0].client as Inngest.Any;
+    for (const app of options.apps) {
+      if (app.client.env !== this.inngest.env) {
+        throw new Error(
+          `All apps must be configured to the same environment. ${app.client.id} is configured to ${app.client.env} but ${this.inngest.id} is configured to ${this.inngest.env}`,
+        );
+      }
+    }
 
     this.options = this.applyDefaults(options);
+
+    this._inngestEnv = this.inngest.env ?? getEnvironmentName();
+
     this.debug = debug("inngest:connect");
 
     this.messageBuffer = new MessageBuffer(this.inngest);
@@ -269,6 +287,18 @@ class WebSocketWorkerConnection implements WorkerConnection {
     this._hashedSigningKey = this.options.signingKey
       ? hashSigningKey(this.options.signingKey)
       : undefined;
+
+    if (
+      this.options.signingKey &&
+      this.options.signingKey.startsWith(
+        InngestBranchEnvironmentSigningKeyPrefix,
+      ) &&
+      !this._inngestEnv
+    ) {
+      throw new Error(
+        "Environment is required when using branch environment signing keys",
+      );
+    }
 
     if (this.options.signingKeyFallback) {
       this._hashedFallbackKey = hashSigningKey(this.options.signingKeyFallback);
@@ -515,9 +545,8 @@ class WebSocketWorkerConnection implements WorkerConnection {
         : {}),
     };
 
-    const envOverride = this.inngest.env;
-    if (envOverride) {
-      headers[headerKeys.Environment] = envOverride;
+    if (this._inngestEnv) {
+      headers[headerKeys.Environment] = this._inngestEnv;
     }
 
     // refactor this to a more universal spot
@@ -543,7 +572,7 @@ class WebSocketWorkerConnection implements WorkerConnection {
       if (resp.status === 401) {
         throw new AuthError(
           `Failed initial API handshake request to ${targetUrl.toString()}${
-            envOverride ? ` (env: ${envOverride})` : ""
+            this._inngestEnv ? ` (env: ${this._inngestEnv})` : ""
           }, ${await resp.text()}`,
           attempt,
         );
@@ -713,7 +742,7 @@ class WebSocketWorkerConnection implements WorkerConnection {
       if (!setupState.sentWorkerConnect) {
         const workerConnectRequestMsg = WorkerConnectRequestData.create({
           connectionId: startResp.connectionId,
-          environment: this.inngest.env || undefined,
+          environment: this._inngestEnv,
           platform: getPlatformName({
             ...allProcessEnv(),
           }),